// src/types/interfaces/course.interface.ts
export interface CourseInterface {
  id: number;
  title: string;
  description: string;
  startDate: string;
  endDate: string;
  duration: string;
  price: string;
  prerequisites: string;
  instructor: string;
  modality: string;
  schedule: string;
  materials: string;
  certification: boolean;
  seats: string;
  location: string;
  category: string;
  imageUrl: string;
}
  